feat: removed jupytyer pre-commit hook
